+++ src/Explore.css
@@ -1,0 +1,5 @@
+.ExploreCardCaption {
+  display: block;
+  color: rgba(0, 0, 255, 0.3);
+  background: rgba(0, 0, 255, 0.3);
+}
